Craft now checks whether a file exists before attempting to delete it.
Resolve #9957

+++ src/controllers/UsersController.php
@@ -1440,7 +1440,7 @@ class UsersController extends Controller
                 'html' => $this->_renderPhotoTemplate($user),
             ]);
         } catch (\Throwable $exception) {
-            if (isset($fileLocation)) {
+            if (isset($fileLocation) && file_exists($fileLocation)) {
                 FileHelper::unlink($fileLocation);
             }
